refactor: BankError, BankMessage 파일 추가
